requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pandas>=1.5.0
pdfplumber>=0.9.0
python-docx>=0.8.11
